requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
asyncpg>=0.27.0
httpx>=0.24.0
pydantic>=1.10.0
python-dotenv>=1.0.0

# FastAPI+uvicorn for app; asyncpg for DB; httpx for provider REST calls; pydantic for validation; python-dotenv for local env loading.